+++ e2e/e2etests/e2etests.go
@@ -96,8 +96,9 @@ const (
 	TestSuiTokenDepositAndCallName           = "sui_token_deposit_and_call"        // #nosec G101: Potential hardcoded credentials (gosec), not a credential
 	TestSuiTokenDepositAndCallRevertName     = "sui_token_deposit_and_call_revert" // #nosec G101: Potential hardcoded credentials (gosec), not a credential
 	TestSuiWithdrawName                      = "sui_withdraw"
+	TestSuiTokenWithdrawName                 = "sui_token_withdraw" // #nosec G101: Potential hardcoded credentials (gosec), not a credential
+	TestSuiWithdrawAndCallName               = "sui_withdraw_and_call"
 	TestSuiWithdrawAndCallRevertWithCallName = "sui_withdraw_and_call_revert_with_call" // #nosec G101: Potential hardcoded credentials (gosec), not a credential
-	TestSuiTokenWithdrawName                 = "sui_token_withdraw"                     // #nosec G101: Potential hardcoded credentials (gosec), not a credential
 
 	/*
 	 Bitcoin tests
@@ -801,12 +802,22 @@ var AllE2ETests = []runner.E2ETest{
 		runner.WithMinimumVersion("v29.0.0"),
 	),
 	runner.NewE2ETest(
+		TestSuiWithdrawAndCallName,
+		"withdraw SUI from ZEVM and call a contract",
+		[]runner.ArgDefinition{
+			{Description: "amount in mist", DefaultValue: "1000000"},
+		},
+		TestSuiWithdrawAndCall,
+		runner.WithMinimumVersion("v30.0.0"),
+	),
+	runner.NewE2ETest(
 		TestSuiWithdrawAndCallRevertWithCallName,
 		"withdraw SUI from ZEVM and call a contract that reverts with a onRevert call",
 		[]runner.ArgDefinition{
 			{Description: "amount in mist", DefaultValue: "1000000"},
 		},
 		TestSuiWithdrawAndCallRevertWithCall,
+		runner.WithMinimumVersion("v30.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiTokenWithdrawName,