+++ zetaclient/chains/sui/signer/signer_tx.go
@@ -2,6 +2,7 @@ package signer
 
 import (
 	"context"
+	"encoding/hex"
 	"fmt"
 	"strconv"
 	"time"
@@ -24,34 +25,71 @@ const (
 	funcIncreaseNonce = "increase_nonce"
 )
 
-// buildWithdraw builds unsigned 'withdraw' transaction using CCTX and Sui RPC
+// outboundDataFromCCTX
+
+// TODO: use these functions in PTB building
+// https://github.com/zeta-chain/node/issues/3741
+//const funcWithdrawImpl = "withdraw_impl"
+//const funcOnCall = "on_call"
+
+// buildWithdrawal builds unsigned withdrawal transaction using CCTX and Sui RPC
 // https://github.com/zeta-chain/protocol-contracts-sui/blob/0245ad3a2eb4001381625070fd76c87c165589b2/sources/gateway.move#L117
-func (s *Signer) buildWithdraw(ctx context.Context, cctx *cctypes.CrossChainTx) (tx models.TxnMetaData, err error) {
-	var (
-		params    = cctx.GetCurrentOutboundParam()
-		nonce     = strconv.FormatUint(params.TssNonce, 10)
-		recipient = params.Receiver
-		amount    = params.Amount.String()
-	)
+func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx) (tx models.TxnMetaData, err error) {
+	params := cctx.GetCurrentOutboundParam()
 
-	// perform basic validation on CCTX
-	coinType, err := validateCCTX(s.Chain().ChainId, cctx)
-	if err != nil {
-		return tx, errors.Wrap(err, "CCTX validation failed")
+	coinType := ""
+
+	// Basic common-sense validation & coin-type determination
+	switch {
+	case params.ReceiverChainId != s.Chain().ChainId:
+		return tx, errors.Errorf("invalid receiver chain id %d", params.ReceiverChainId)
+	case cctx.ProtocolContractVersion != cctypes.ProtocolContractVersion_V2:
+		return tx, errors.Errorf("invalid protocol version %q", cctx.ProtocolContractVersion)
+	case cctx.InboundParams == nil:
+		return tx, errors.New("inbound params are nil")
+	case params.CoinType == coin.CoinType_Gas:
+		coinType = string(sui.SUI)
+	case params.CoinType == coin.CoinType_ERC20:
+		// NOTE: 0x prefix is required for coin type other than SUI
+		coinType = "0x" + cctx.InboundParams.Asset
+	default:
+		return tx, errors.Errorf("unsupported coin type %q", params.CoinType.String())
 	}
 
-	// get gas budget from CCTX
-	gasBudget, err := gasBudgetFromCCTX(cctx)
+	// Gas budget is gas limit * gas price
+	gasPrice, err := strconv.ParseUint(params.GasPrice, 10, 64)
 	if err != nil {
-		return tx, errors.Wrap(err, "unable to get gas budget")
+		return tx, errors.Wrap(err, "unable to parse gas price")
 	}
+	gasBudget := strconv.FormatUint(gasPrice*params.CallOptions.GasLimit, 10)
 
+	// Retrieve withdraw cap ID
 	withdrawCapID, err := s.getWithdrawCapIDCached(ctx)
 	if err != nil {
 		return tx, errors.Wrap(err, "unable to get withdraw cap ID")
 	}
 
-	// build withdraw tx
+	// build tx depending on the type of transaction
+	if cctx.IsWithdrawAndCall() {
+		return s.buildWithdrawAndCallTx(ctx, params, coinType, gasBudget, withdrawCapID, cctx.RelayedMessage)
+	}
+	return s.buildWithdrawTx(ctx, params, coinType, gasBudget, withdrawCapID)
+}
+
+// buildWithdrawTx builds unsigned withdraw transaction
+func (s *Signer) buildWithdrawTx(
+	ctx context.Context,
+	params *cctypes.OutboundParams,
+	coinType,
+	gasBudget,
+	withdrawCapID string,
+) (models.TxnMetaData, error) {
+	var (
+		nonce     = strconv.FormatUint(params.TssNonce, 10)
+		recipient = params.Receiver
+		amount    = params.Amount.String()
+	)
+
 	req := models.MoveCallRequest{
 		Signer:          s.TSS().PubKey().AddressSui(),
 		PackageObjectId: s.gateway.PackageID(),
@@ -65,85 +103,75 @@ func (s *Signer) buildWithdraw(ctx context.Context, cctx *cctypes.CrossChainTx) 
 	return s.client.MoveCall(ctx, req)
 }
 
-// buildExecute builds unsigned 'execute' transaction using CCTX and Sui RPC
-//
-// TODO: replace it with a real implementation
-// this function is a fake implementation for testing the tx cancellation
-func (s *Signer) buildExecute(
+// buildWithdrawAndCallTx builds unsigned withdrawAndCall
+// a withdrawAndCall is a PTB transaction that contains a withdraw_impl call and a on_call call
+func (s *Signer) buildWithdrawAndCallTx(
 	ctx context.Context,
-	cctx *cctypes.CrossChainTx,
-) (tx models.TxnMetaData, err error) {
-	var (
-		params    = cctx.GetCurrentOutboundParam()
-		nonce     = strconv.FormatUint(params.TssNonce, 10)
-		recipient = params.Receiver
-		amount    = params.Amount.String()
+	params *cctypes.OutboundParams,
+	coinType,
+	gasBudget,
+	withdrawCapID,
+	payload string,
+) (models.TxnMetaData, error) {
+	// decode and parse the payload to object the on_call arguments
+	payloadBytes, err := hex.DecodeString(payload)
+	if err != nil {
+		return models.TxnMetaData{}, errors.Wrap(err, "unable to decode payload hex bytes")
+	}
+
+	var cp sui.CallPayload
+	if err := cp.UnpackABI(payloadBytes); err != nil {
+		return models.TxnMetaData{}, errors.Wrap(err, "unable to parse withdrawAndCall payload")
+	}
+
+	// Note: logs not formatted in standard, it's a temporary log
+	s.Logger().Std.Info().Msgf(
+		"WithdrawAndCall called with type arguments %v, object IDs %v, message %v",
+		cp.TypeArgs,
+		cp.ObjectIDs,
+		cp.Message,
 	)
 
-	// perform basic validation on CCTX
-	coinType, err := validateCCTX(s.Chain().ChainId, cctx)
-	if err != nil {
-		return tx, errors.Wrap(err, "CCTX validation failed")
-	}
+	// keep lint quiet without using _ in params
+	_ = ctx
+	_ = params
+	_ = coinType
+	_ = gasBudget
+	_ = withdrawCapID
 
-	// use a fake coin type to fail the withdraw tx
-	// this will force the withdraw to be cancelled
-	if cctx.InboundParams.CoinType == coin.CoinType_Gas {
-		coinType = coinType[0:64] + "::fake::FAKE"
-	} else {
-		coinType = coinType[0:66] + "::fake::FAKE"
-	}
+	// TODO: check all object IDs are share object here
+	// https://github.com/zeta-chain/node/issues/3755
 
-	// get gas budget from CCTX
-	gasBudget, err := gasBudgetFromCCTX(cctx)
-	if err != nil {
-		return tx, errors.Wrap(err, "unable to get gas budget")
-	}
+	// TODO: build PTB here
+	// https://github.com/zeta-chain/node/issues/3741
 
-	withdrawCapID, err := s.getWithdrawCapIDCached(ctx)
-	if err != nil {
-		return tx, errors.Wrap(err, "unable to get withdraw cap ID")
-	}
-
-	// build a withdraw tx as a fake 'execute' because we don't have a real one yet
-	req := models.MoveCallRequest{
-		Signer:          s.TSS().PubKey().AddressSui(),
-		PackageObjectId: s.gateway.PackageID(),
-		Module:          s.gateway.Module(),
-		Function:        funcWithdraw, // TODO: change to funcExecute
-		TypeArguments:   []any{coinType},
-		Arguments:       []any{s.gateway.ObjectID(), amount, nonce, recipient, gasBudget, withdrawCapID},
-		GasBudget:       gasBudget,
-	}
-
-	return s.client.MoveCall(ctx, req)
+	return models.TxnMetaData{}, errors.New("not implemented")
 }
 
-// buildIncreaseNonce builds unsigned 'increase_nonce' transaction using CCTX and Sui RPC
-func (s *Signer) buildIncreaseNonce(
+// createCancelTxBuilder creates a cancel tx builder for given CCTX
+// The tx cancellation is done by calling the 'increase_nonce' function on the gateway
+func (s *Signer) createCancelTxBuilder(
 	ctx context.Context,
 	cctx *cctypes.CrossChainTx,
-) (tx models.TxnMetaData, err error) {
-	nonce := strconv.FormatUint(cctx.GetCurrentOutboundParam().TssNonce, 10)
-
-	// perform basic validation on CCTX
-	_, err = validateCCTX(s.Chain().ChainId, cctx)
-	if err != nil {
-		return tx, errors.Wrap(err, "CCTX validation failed")
-	}
+	zetaHeight uint64,
+) (txBuilder, error) {
+	var (
+		params = cctx.GetCurrentOutboundParam()
+		nonce  = strconv.FormatUint(params.TssNonce, 10)
+	)
 
 	// get gas budget from CCTX
 	gasBudget, err := gasBudgetFromCCTX(cctx)
 	if err != nil {
-		return tx, errors.Wrap(err, "unable to get gas budget")
+		return nil, errors.Wrap(err, "unable to get gas budget")
 	}
 
+	// retrieve withdraw cap ID
 	withdrawCapID, err := s.getWithdrawCapIDCached(ctx)
 	if err != nil {
-		return tx, errors.Wrap(err, "unable to get withdraw cap ID")
+		return nil, errors.Wrap(err, "unable to get withdraw cap ID")
 	}
 
-	// build increase nonce tx
 	req := models.MoveCallRequest{
 		Signer:          s.TSS().PubKey().AddressSui(),
 		PackageObjectId: s.gateway.PackageID(),
@@ -154,38 +182,40 @@ func (s *Signer) buildIncreaseNonce(
 		GasBudget:       gasBudget,
 	}
 
-	return s.client.MoveCall(ctx, req)
-}
-
-// buildExecuteWithCancelTxBuilder builds both unsigned 'execute' and a cancel tx
-func (s *Signer) buildExecuteWithCancelTxBuilder(
-	ctx context.Context,
-	cctx *cctypes.CrossChainTx,
-	zetaHeight uint64,
-) (models.TxnMetaData, txBuilder, error) {
-	nonce := cctx.GetCurrentOutboundParam().TssNonce
-
-	tx, err := s.buildExecute(ctx, cctx)
-	if err != nil {
-		return tx, nil, errors.Wrap(err, "unable to build execute tx")
-	}
-
 	// tx builder for cancel tx
-	cancelTxBuilder := func() (models.TxnMetaData, string, error) {
-		txCancel, err := s.buildIncreaseNonce(ctx, cctx)
+	// delay the 'MoveCall' to the last moment helps to avoid gateway object version mismatch
+	return func() (models.TxnMetaData, string, error) {
+		txCancel, err := s.client.MoveCall(ctx, req)
 		if err != nil {
 			return models.TxnMetaData{}, "", errors.Wrap(err, "unable to build cancel tx")
 		}
 
-		sigCancel, err := s.signTx(ctx, txCancel, zetaHeight, nonce)
+		sigCancel, err := s.signTx(ctx, txCancel, zetaHeight, params.TssNonce)
 		if err != nil {
 			return models.TxnMetaData{}, "", errors.Wrap(err, "unable to sign cancel tx")
 		}
-
 		return txCancel, sigCancel, nil
+	}, nil
+}
+
+// broadcast attaches signature to tx and broadcasts it to Sui network. Returns tx digest.
+func (s *Signer) broadcast(ctx context.Context, tx models.TxnMetaData, sig [65]byte) (string, error) {
+	sigBase64, err := sui.SerializeSignatureECDSA(sig, s.TSS().PubKey().AsECDSA())
+	if err != nil {
+		return "", errors.Wrap(err, "unable to serialize signature")
 	}
 
-	return tx, cancelTxBuilder, nil
+	req := models.SuiExecuteTransactionBlockRequest{
+		TxBytes:   tx.TxBytes,
+		Signature: []string{sigBase64},
+	}
+
+	res, err := s.client.SuiExecuteTransactionBlock(ctx, req)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to execute tx block")
+	}
+
+	return res.Digest, nil
 }
 
 // broadcastWithCancelTx attaches signature to tx and broadcasts it to Sui network. Returns tx digest.
@@ -259,29 +289,6 @@ func (s *Signer) broadcastWithCancelTx(
 	logger.Info().Str(logs.FieldTx, res.Digest).Msg("Executed sui cancel tx block")
 
 	return res.Digest, nil
-}
-
-// validateCCTX performs basic common-sense validation on CCTX and determines coin type
-func validateCCTX(signerChainID int64, cctx *cctypes.CrossChainTx) (coinType string, err error) {
-	params := cctx.GetCurrentOutboundParam()
-
-	switch {
-	case params.ReceiverChainId != signerChainID:
-		return "", errors.Errorf("invalid receiver chain id %d", params.ReceiverChainId)
-	case cctx.ProtocolContractVersion != cctypes.ProtocolContractVersion_V2:
-		return "", errors.Errorf("invalid protocol version %q", cctx.ProtocolContractVersion)
-	case cctx.InboundParams == nil:
-		return "", errors.New("inbound params are nil")
-	case cctx.InboundParams.CoinType == coin.CoinType_Gas:
-		coinType = string(sui.SUI)
-	case cctx.InboundParams.CoinType == coin.CoinType_ERC20:
-		// NOTE: 0x prefix is required for coin type other than SUI
-		coinType = "0x" + cctx.InboundParams.Asset
-	default:
-		return "", errors.Errorf("unsupported coin type %q", cctx.InboundParams.CoinType.String())
-	}
-
-	return coinType, nil
 }
 
 // gasBudgetFromCCTX returns gas budget from CCTX